ST-1135: Working changes
